add vue-router and move App.vue to components/hackmap
