+++ package.json
@@ -17,7 +17,8 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^0.16.1",
-    "vue": "^2.3.3"
+    "vue": "^2.3.3",
+    "vue-router": "^2.7.0"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.2",
